requirements.txt
aiofiles
aiohttp
pycryptodome